+++ src/spark/quest/SimpleQuest.java
@@ -17,7 +17,7 @@ public class SimpleQuest extends Quest
      * @param JumpA The transition ID for option A
      * @param JumpB The transition ID for option A
      */
-    public SimpleQuest(long ID, String Desc, String OptionA, String OptionB, long JumpA, long JumpB)
+    public SimpleQuest(int ID, String Desc, String OptionA, String OptionB, int JumpA, int JumpB)
     {
         super(ID, Desc, OptionA, OptionB, JumpA, JumpB);
     }
